--- src/contexts/AuthContext.js
@@ -1,50 +1,0 @@
-import { createContext, useState, useEffect } from 'react';
-import { currentUser } from '../data/userData';
-
-const AuthContext = createContext();
-
-export const AuthProvider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    // 페이지 새로고침 시 로컬 스토리지에서 사용자 정보 확인
-    const savedUser = localStorage.getItem('user');
-    if (savedUser) {
-      setUser(JSON.parse(savedUser));
-    }
-    setLoading(false);
-  }, []);
-
-  const login = async (username, password) => {
-    // API 호출을 통해 로그인 처리
-    if (username === 'admin' && password === 'password') {
-      setUser(currentUser);
-      localStorage.setItem('user', JSON.stringify(currentUser));
-      return currentUser;
-    } else {
-      throw new Error('Invalid credentials');
-    }
-  };
-
-  const logout = () => {
-    setUser(null);
-    localStorage.removeItem('user');
-  };
-
-  const value = {
-    user,
-    login,
-    logout,
-    loading,
-    isAuthenticated: !!user
-  };
-
-  return (
-    <AuthContext.Provider value={value}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export default AuthContext;
